blockchain: Single out treasury policy test.
This is purely a code move.
